+++ crates/ff_kinetics/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to this crate will be documented in this file.
 
+## development:
+- another speedup for stochastic simulations.
+
 ## [0.3.0] - 2026-01-13
 ### Added
 - rate_tree.rs: logarithmic neighbor selection in stochastic simulations (huge speedup).